+++ dvc/testing/conftest.py
@@ -1,0 +1,1 @@
+from .fixtures import *  # noqa, pylint: disable=wildcard-import
