+++ dvc/testing/path_info.py
