+++ setup.cfg
@@ -158,6 +158,8 @@ tests =
 
 [options.packages.find]
 exclude =
+    dvc.testing
+    dvc.testing.*
     tests
     tests.*
