+++ tests/dir_helpers.py
@@ -47,23 +47,14 @@ from global repo template to creating everything inplace, which:
 
 import os
 import pathlib
-import sys
-from contextlib import contextmanager
-from functools import partialmethod
 from textwrap import dedent
 
 import pytest
-from funcy import lmap, retry
 
 from dvc.logger import disable_other_loggers
-from dvc.utils import serialize
-from dvc.utils.fs import makedirs
+from dvc.testing.tmp_dir import git_init
 
 __all__ = [
-    "make_tmp_dir",
-    "tmp_dir",
-    "scm",
-    "dvc",
     "run_copy",
     "run_head",
     "erepo_dir",
@@ -76,299 +67,6 @@ __all__ = [
 
 # see https://github.com/iterative/dvc/issues/3167
 disable_other_loggers()
-
-
-class TmpDir(pathlib.Path):
-    scheme = "local"
-
-    @property
-    def fs_path(self):
-        return str(self)
-
-    def __new__(cls, *args, **kwargs):
-        if cls is TmpDir:
-            cls = (  # pylint: disable=self-cls-assignment
-                WindowsTmpDir if os.name == "nt" else PosixTmpDir
-            )
-        # init parameter and `_init` method has been removed in Python 3.10.
-        kw = {"init": False} if sys.version_info < (3, 10) else {}
-        self = cls._from_parts(  # pylint: disable=unexpected-keyword-arg
-            args, **kw
-        )
-        if not self._flavour.is_supported:
-            raise NotImplementedError(
-                f"cannot instantiate {cls.__name__!r} on your system"
-            )
-        if sys.version_info < (3, 10):
-            self._init()  # pylint: disable=no-member
-        return self
-
-    def init(self, *, scm=False, dvc=False, subdir=False):
-        from scmrepo.git import Git
-
-        from dvc.repo import Repo
-
-        assert not scm or not hasattr(self, "scm")
-        assert not dvc or not hasattr(self, "dvc")
-
-        str_path = os.fspath(self)
-
-        if scm:
-            git_init(str_path)
-        if dvc:
-            self.dvc = Repo.init(
-                str_path,
-                no_scm=not scm and not hasattr(self, "scm"),
-                subdir=subdir,
-            )
-        if scm:
-            self.scm = self.dvc.scm if hasattr(self, "dvc") else Git(str_path)
-        if dvc and hasattr(self, "scm"):
-            self.scm.commit("init dvc")
-
-    def close(self):
-        if hasattr(self, "scm"):
-            self.scm.close()
-        if hasattr(self, "dvc"):
-            self.dvc.close()
-
-    def _require(self, name):
-        if not hasattr(self, name):
-            raise TypeError(
-                "Can't use {name} for this temporary dir. "
-                'Did you forget to use "{name}" fixture?'.format(name=name)
-            )
-
-    # Bootstrapping methods
-    def gen(self, struct, text=""):
-        if isinstance(struct, (str, bytes, pathlib.PurePath)):
-            struct = {struct: text}
-
-        return self._gen(struct)
-
-    def _gen(self, struct, prefix=None):
-        paths = []
-        for name, contents in struct.items():
-            path = (prefix or self) / name
-
-            if isinstance(contents, dict):
-                if not contents:
-                    makedirs(path, exist_ok=True)
-                else:
-                    self._gen(contents, prefix=path)
-            else:
-                makedirs(path.parent, exist_ok=True)
-                if isinstance(contents, bytes):
-                    path.write_bytes(contents)
-                else:
-                    path.write_text(contents, encoding="utf-8")
-            paths.append(path)
-        return paths
-
-    def dvc_gen(self, struct, text="", commit=None):
-        paths = self.gen(struct, text)
-        return self.dvc_add(paths, commit=commit)
-
-    def scm_gen(self, struct, text="", commit=None):
-        paths = self.gen(struct, text)
-        return self.scm_add(paths, commit=commit)
-
-    def commit(self, output_paths, msg):
-        def to_gitignore(stage_path):
-            from dvc.scm import Git
-
-            return os.path.join(os.path.dirname(stage_path), Git.GITIGNORE)
-
-        gitignores = [
-            to_gitignore(s)
-            for s in output_paths
-            if os.path.exists(to_gitignore(s))
-        ]
-        return self.scm_add(output_paths + gitignores, commit=msg)
-
-    def dvc_add(self, filenames, commit=None):
-        self._require("dvc")
-        filenames = _coerce_filenames(filenames)
-
-        stages = self.dvc.add(filenames)
-        if commit:
-            self.commit([s.path for s in stages], msg=commit)
-        return stages
-
-    def scm_add(self, filenames, commit=None):
-        self._require("scm")
-        filenames = _coerce_filenames(filenames)
-        self.scm.add(filenames)
-        if commit:
-            self.scm.commit(commit)
-
-    def add_remote(
-        self, *, url=None, config=None, name="upstream", default=True
-    ):
-        self._require("dvc")
-
-        assert bool(url) ^ bool(config)
-
-        if url:
-            config = {"url": url}
-
-        with self.dvc.config.edit() as conf:
-            conf["remote"][name] = config
-            if default:
-                conf["core"]["remote"] = name
-
-        if hasattr(self, "scm"):
-            self.scm.add(self.dvc.config.files["repo"])
-            self.scm.commit(f"add '{name}' remote")
-
-        return url or config["url"]
-
-    # contexts
-    @contextmanager
-    def chdir(self):
-        old = os.getcwd()
-        try:
-            os.chdir(self)
-            yield
-        finally:
-            os.chdir(old)
-
-    @contextmanager
-    def branch(self, name, new=False):
-        self._require("scm")
-        old = self.scm.active_branch()
-        try:
-            self.scm.checkout(name, create_new=new)
-            yield
-        finally:
-            self.scm.checkout(old)
-
-    def read_text(self, *args, **kwargs):  # pylint: disable=signature-differs
-        # NOTE: on windows we'll get PermissionError instead of
-        # IsADirectoryError when we try to `open` a directory, so we can't
-        # rely on exception flow control
-        if self.is_dir():
-            return {
-                path.name: path.read_text(*args, **kwargs)
-                for path in self.iterdir()
-            }
-        return super().read_text(*args, encoding="utf-8", **kwargs)
-
-    def hash_to_path(self, hash_):
-        return str(self / hash_[0:2] / hash_[2:])
-
-    def dump(self, *args, **kwargs):
-        return serialize.DUMPERS[self.suffix](self, *args, **kwargs)
-
-    def parse(self, *args, **kwargs):
-        return serialize.LOADERS[self.suffix](self, *args, **kwargs)
-
-    def modify(self, *args, **kwargs):
-        return serialize.MODIFIERS[self.suffix](self, *args, **kwargs)
-
-    load_yaml = partialmethod(serialize.load_yaml)
-    dump_yaml = partialmethod(serialize.dump_yaml)
-    load_json = partialmethod(serialize.load_json)
-    dump_json = partialmethod(serialize.dump_json)
-    load_toml = partialmethod(serialize.load_toml)
-    dump_toml = partialmethod(serialize.dump_toml)
-
-
-def _coerce_filenames(filenames):
-    if isinstance(filenames, (str, bytes, pathlib.PurePath)):
-        filenames = [filenames]
-    return lmap(os.fspath, filenames)
-
-
-class WindowsTmpDir(TmpDir, pathlib.PureWindowsPath):
-    pass
-
-
-class PosixTmpDir(TmpDir, pathlib.PurePosixPath):
-    pass
-
-
-CACHE = {}
-
-
-def _fs_copy(src, dst, ignore=None):
-    import shutil
-
-    if os.path.isdir(src):
-        shutil.copytree(src, dst, ignore=ignore)
-    else:
-        shutil.copy2(src, dst)
-
-
-@pytest.fixture(scope="session")
-def make_tmp_dir(tmp_path_factory, request, worker_id):
-    def make(name, *, scm=False, dvc=False, subdir=False):
-        from shutil import ignore_patterns
-
-        from scmrepo.git import Git
-
-        from dvc.repo import Repo
-
-        cache = CACHE.get((scm, dvc, subdir))
-        if not cache:
-            cache = tmp_path_factory.mktemp("dvc-test-cache" + worker_id)
-            TmpDir(cache).init(scm=scm, dvc=dvc, subdir=subdir)
-            CACHE[(scm, dvc, subdir)] = os.fspath(cache)
-        path = tmp_path_factory.mktemp(name) if isinstance(name, str) else name
-
-        # ignore sqlite files from .dvc/tmp. We might not be closing the cache
-        # connection resulting in PermissionErrors in Windows.
-        ignore = ignore_patterns("cache.db*")
-        for entry in os.listdir(cache):
-            # shutil.copytree's dirs_exist_ok is only available in >=3.8
-            _fs_copy(
-                os.path.join(cache, entry),
-                os.path.join(path, entry),
-                ignore=ignore,
-            )
-        new_dir = TmpDir(path)
-        str_path = os.fspath(new_dir)
-        if dvc:
-            new_dir.dvc = Repo(str_path)
-        if scm:
-            new_dir.scm = (
-                new_dir.dvc.scm if hasattr(new_dir, "dvc") else Git(str_path)
-            )
-        request.addfinalizer(new_dir.close)
-        return new_dir
-
-    return make
-
-
-@pytest.fixture
-def tmp_dir(tmp_path, make_tmp_dir, request, monkeypatch):
-    monkeypatch.chdir(tmp_path)
-    fixtures = request.fixturenames
-    return make_tmp_dir(tmp_path, scm="scm" in fixtures, dvc="dvc" in fixtures)
-
-
-@pytest.fixture
-def scm(tmp_dir):
-    return tmp_dir.scm
-
-
-@pytest.fixture
-def dvc(tmp_dir):
-    with tmp_dir.dvc as _dvc:
-        yield _dvc
-
-
-def git_init(path):
-    from git import Repo
-    from git.exc import GitCommandNotFound
-
-    # NOTE: handles EAGAIN error on BSD systems (osx in our case).
-    # Otherwise when running tests you might get this exception:
-    #
-    #    GitCommandNotFound: Cmd('git') not found due to:
-    #        OSError('[Errno 35] Resource temporarily unavailable')
-    git = retry(5, GitCommandNotFound)(Repo.init)(path)
-    git.close()
 
 
 @pytest.fixture
@@ -443,14 +141,19 @@ def git_dir(make_tmp_dir):
     return path
 
 
+class GitRemote:
+    def __init__(self, tmp_dir, name, url):
+        self.tmp_dir = tmp_dir
+        self.remote = name
+        self.url = url
+
+
 @pytest.fixture
 def git_upstream(tmp_dir, erepo_dir, git_dir, request):
     remote = erepo_dir if "dvc" in request.fixturenames else git_dir
     url = "file://{}".format(remote.resolve().as_posix())
     tmp_dir.scm.gitpython.repo.create_remote("upstream", url)
-    remote.remote = "upstream"
-    remote.url = url
-    return remote
+    return GitRemote(remote, "upstream", url)
 
 
 @pytest.fixture
@@ -458,6 +161,4 @@ def git_downstream(tmp_dir, erepo_dir, git_dir, request):
     remote = erepo_dir if "dvc" in request.fixturenames else git_dir
     url = "file://{}".format(tmp_dir.resolve().as_posix())
     remote.scm.gitpython.repo.create_remote("upstream", url)
-    remote.remote = "upstream"
-    remote.url = url
-    return remote
+    return GitRemote(remote, "upstream", url)
