+++ tests/func/experiments/test_remove.py
@@ -107,12 +107,14 @@ def test_remove_remote(tmp_dir, scm, dvc, exp_stage, git_upstream, use_url):
         ref_info = first(exp_refs_by_rev(scm, exp))
         ref_info_list.append(ref_info)
         dvc.experiments.push(remote, ref_info.name)
-        assert git_upstream.scm.get_ref(str(ref_info)) == exp
+        assert git_upstream.tmp_dir.scm.get_ref(str(ref_info)) == exp
 
     dvc.experiments.remove(
         remote=remote, exp_names=[str(ref_info_list[0]), ref_info_list[1].name]
     )
 
-    assert git_upstream.scm.get_ref(str(ref_info_list[0])) is None
-    assert git_upstream.scm.get_ref(str(ref_info_list[1])) is None
-    assert git_upstream.scm.get_ref(str(ref_info_list[2])) == exp_list[2]
+    assert git_upstream.tmp_dir.scm.get_ref(str(ref_info_list[0])) is None
+    assert git_upstream.tmp_dir.scm.get_ref(str(ref_info_list[1])) is None
+    assert (
+        git_upstream.tmp_dir.scm.get_ref(str(ref_info_list[2])) == exp_list[2]
+    )
