+++ tests/remotes/oss.py
@@ -4,10 +4,9 @@ import uuid
 
 import pytest
 
+from dvc.testing.cloud import Cloud
+from dvc.testing.path_info import CloudURLInfo
 from dvc.utils import env2bool
-
-from .base import Base
-from .path_info import CloudURLInfo
 
 TEST_OSS_REPO_BUCKET = "dvc-test-github"
 EMULATOR_OSS_ENDPOINT = "127.0.0.1:{port}"
@@ -15,7 +14,7 @@ EMULATOR_OSS_ACCESS_KEY_ID = "AccessKeyID"
 EMULATOR_OSS_ACCESS_KEY_SECRET = "AccessKeySecret"
 
 
-class OSS(Base, CloudURLInfo):
+class OSS(Cloud, CloudURLInfo):
 
     IS_OBJECT_STORAGE = True
 
@@ -35,6 +34,33 @@ class OSS(Base, CloudURLInfo):
             return True
 
         return False
+
+    @property
+    def config(self):
+        return {
+            "url": self.url,
+            "oss_key_id": os.environ.get("OSS_ACCESS_KEY_ID"),
+            "oss_key_secret": os.environ.get("OSS_ACCESS_KEY_SECRET"),
+            "oss_endpoint": os.environ.get("OSS_ENDPOINT"),
+        }
+
+    def is_file(self):
+        raise NotImplementedError
+
+    def is_dir(self):
+        raise NotImplementedError
+
+    def exists(self):
+        raise NotImplementedError
+
+    def mkdir(self, mode=0o777, parents=False, exist_ok=False):
+        raise NotImplementedError
+
+    def write_bytes(self, contents):
+        raise NotImplementedError
+
+    def read_bytes(self):
+        raise NotImplementedError
 
 
 @pytest.fixture(scope="session")
@@ -63,24 +89,36 @@ def oss_server(test_config, docker_compose, docker_services):
 
 
 @pytest.fixture
-def oss(real_oss):
-    import oss2
+def make_oss(real_oss):
+    def _make_oss():
+        import oss2
 
-    ret = real_oss
+        ret = real_oss
 
-    auth = oss2.Auth(ret.config["oss_key_id"], ret.config["oss_key_secret"])
-    bucket = oss2.Bucket(
-        auth, ret.config["oss_endpoint"], TEST_OSS_REPO_BUCKET
-    )
-    try:
-        bucket.get_bucket_info()
-    except oss2.exceptions.NoSuchBucket:
-        bucket.create_bucket(
-            oss2.BUCKET_ACL_PUBLIC_READ,
-            oss2.models.BucketCreateConfig(oss2.BUCKET_STORAGE_CLASS_STANDARD),
+        auth = oss2.Auth(
+            ret.config["oss_key_id"], ret.config["oss_key_secret"]
         )
+        bucket = oss2.Bucket(
+            auth, ret.config["oss_endpoint"], TEST_OSS_REPO_BUCKET
+        )
+        try:
+            bucket.get_bucket_info()
+        except oss2.exceptions.NoSuchBucket:
+            bucket.create_bucket(
+                oss2.BUCKET_ACL_PUBLIC_READ,
+                oss2.models.BucketCreateConfig(
+                    oss2.BUCKET_STORAGE_CLASS_STANDARD
+                ),
+            )
 
-    return ret
+        return ret
+
+    return _make_oss
+
+
+@pytest.fixture
+def oss(make_oss):
+    return make_oss()
 
 
 @pytest.fixture
@@ -90,11 +128,4 @@ def real_oss(test_config):
         pytest.skip("no real OSS")
 
     url = OSS.get_url()
-    ret = OSS(url)
-    ret.config = {
-        "url": url,
-        "oss_key_id": os.environ.get("OSS_ACCESS_KEY_ID"),
-        "oss_key_secret": os.environ.get("OSS_ACCESS_KEY_SECRET"),
-        "oss_endpoint": os.environ.get("OSS_ENDPOINT"),
-    }
-    return ret
+    return OSS(url)
